Moving validateBody middleware into its own file
