+++ src/middlewares/validateBlog.middleware.ts
@@ -1,21 +1,7 @@
 import { Request, Response, NextFunction } from 'express';
-import { z, ZodType } from 'zod';
+import { z } from 'zod';
 import { blogTextSchema, imageFileSchema } from '../validations/blogSchema';
 import { formatZodError } from '../utils/formatZodError';
-
-export function validateBody<T extends ZodType>(schema: T) {
-  return (req: Request, res: Response, next: NextFunction) => {
-    const result = schema.safeParse(req.body);
-    if (!result.success) {
-      const simplifiedErrors = formatZodError(result.error);
-      return res.status(400).json({
-        error: 'Validation failed',
-        details: simplifiedErrors,
-      });
-    }
-    next();
-  };
-}
 
 export const validateBlog = (req: Request, res: Response, next: NextFunction) => {
   if (!req.user) {
