+++ src/routes/blog.route.ts
@@ -5,7 +5,8 @@ import {
   handleGetBlogList,
 } from '../controllers/blog.controller';
 import { handleAddComment, handleUpdateComment } from '../controllers/comment.controller';
-import { validateBlog, validateBody } from '../middlewares/validateBlog.middleware';
+import { validateBlog } from '../middlewares/validateBlog.middleware';
+import { validateBody } from '../middlewares/validateBody.middleware';
 import { commentSchema } from '../validations/commentSchema';
 import { upload } from '../middlewares/multer.middleware';
 import { handleGetBlogLikes, handleBlogLikeStatus } from '../controllers/blogLike.controller';
